+++ tests/test_retrieval.py
@@ -34,22 +34,22 @@ class TestRetrieve:
             result["4"],
             (
                 596488,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-91a08"',
-                '"68bf6260-91a08"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-91a08"',
+                '"69a8f210-91a08"',
                 True,
                 True,
                 None,
                 200,
-                5,
+                12,
             ),
         )
         check.equal(
             result["5"],
             (
                 8461,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-210d"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-210d"',
                 "914bd375",
                 True,
                 True,
@@ -62,36 +62,36 @@ class TestRetrieve:
             result["6"],
             (
                 26084,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-65e4"',
-                '"68bf6260-65e4"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-65e4"',
+                '"69a8f210-65e4"',
                 None,
                 True,
                 None,
                 200,
-                5,
+                12,
             ),
         )
         check.equal(
             result["7"],
             (
                 81920,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-14000"',
-                '"68bf6260-14000"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-14000"',
+                '"69a8f210-14000"',
                 True,
                 True,
                 None,
                 200,
-                5,
+                12,
             ),
         )
         check.equal(
             result["8"],
             (
                 20984,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-51f8"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-51f8"',
                 "08157f501860340ef6e102e9037d1da2",
                 True,
                 True,
@@ -104,42 +104,42 @@ class TestRetrieve:
             result["9"],
             (
                 2017,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-7e1"',
-                '"68bf6260-7e1"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-7e1"',
+                '"69a8f210-7e1"',
                 None,
                 None,
                 None,
                 200,
-                5,
+                12,
             ),
         )
         check.equal(
             result["10"],
             (
                 2017,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-7e1"',
-                '"68bf6260-7e1"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-7e1"',
+                '"69a8f210-7e1"',
                 None,
                 False,
                 None,
                 200,
-                5,
+                12,
             ),
         )
         check.equal(
             result["11"],
             (
                 2017,
-                "Mon, 08 Sep 2025 23:10:24 GMT",
-                '"68bf6260-7e1"',
-                '"68bf6260-7e1"',
+                "Thu, 05 Mar 2026 03:01:36 GMT",
+                '"69a8f210-7e1"',
+                '"69a8f210-7e1"',
                 False,
                 False,
                 None,
                 200,
-                5,
+                12,
             ),
         )
         check.equal(
@@ -211,6 +211,28 @@ class TestRetrieve:
                 True,
                 None,
                 200,
-                7,
+                100,
             ),
         )
+
+    def test_retrieval_status_10_etag_match(self):
+        """Status 10: Server returns an ETag that matches our existing_hash."""
+        url = "https://data.humdata.org/dataset/ea1d4259-040c-40bd-b1b3-eee9caadf9d3/resource/12b59516-5839-48ce-a13e-3e6d5455dc43/download/pcn_children_under_five_2020_geotiff.zip"
+        retrieval_obj = Retrieval("test", {"data.humdata.org"})
+
+        # 1. Fetch once to retrieve the live server's ETag
+        urls_initial = [("1", "id_1", url, "zip", None, None, None, "N")]
+        res_initial = retrieval_obj.retrieve(urls_initial)
+
+        # Index 2 is the ETag!
+        live_etag = res_initial["id_1"][2]
+
+        # Ensure the server actually provided an ETag before proceeding
+        assert live_etag is not None, "Server did not return an ETag."
+
+        # 2. Fetch again, passing the valid ETag as our existing_hash
+        urls_match = [("1", "id_1", url, "zip", live_etag, None, None, "N")]
+        res_match = retrieval_obj.retrieve(urls_match)
+
+        # Now it will correctly match and short-circuit to 10
+        check.equal(res_match["id_1"][8], 10)